+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * DateBetween增加可选是否取绝对值选项构造（issue#IETE0@gitee）
 * 加入Rythm模板引擎工具类
 * cron模块中增加方法支持获取Task和CronPattern（感谢@Γ平淡ㄎ）
+* HttpResponse中增加个体Cookie方法
 
 ### Bug修复
 * NumberUtil中针对Double重载方法，避免传入包装类型引起的歧义